Provide feedback when authentication fails
Add a `with_authentication` wrapper that executes a function after
ensuring that an API token is available. If an authentication error
occurs, a helpful message is printed to standard error, informing the
user how to proceed.

All existing calls to `CogApi.authenticate/1` were converted to use this
wrapper, and tests for the wrapper logic are in place.

Fixes https://github.com/operable/cog/issues/371