Use the HTTP status code to determine success for bootstrapping
When you bootstrap a system that is already bootstrapped you get a 423
back, which is technically an error. This would cause cogctl to exit
with a status of 1. We want bootstrapping to be idempotent, so now if we
get a 423, we just let the user know that the system is already
bootstrapped and exit with a 0.