Dockerized CI for cogctl tests
Our tests are currently in a state of flux. Previously, _all_ tests were
run against a real Cog server. Recently, however, tests are being
migrated to an ExVCR-based flow which mocks out the network
interactions, and thus run much faster. This commit adds an ExUnit
tag (`:external`) to all test modules that operate against the real Cog
server, allowing us to split the long-running tests apart from the much
quicker ones. The pipeline reflects this, with a **Unit Tests** step
that runs within a single Docker container, and an **Integration Tests**
step, which runs in a `docker-compose` network with a Postgres and a Cog
server. This division (and the way it is implemented) ensures that
future migrations of tests to the ExVCR-based flow will be automatically
reflected in the CI pipeline execution without further intervention.

Additionally, a new **Escript** step is added, in parallel with these
other steps, just to ensure that we can, in fact, generate a `cogctl`
executable; this wasn't ever tested before.

Before all these pipeline steps, however, we do a build of the `cogctl`
Dockerfile. The Dockerfile takes advantage of Docker caching to retrieve
and compile all the dependencies, both for `MIX_ENV=test` (for both the
test steps), as well as for `MIX_ENV=prod` (for the escript
generation). Dependencies are not as quick to change as core code, and
so caching can dramatically shorten build time, particularly on
subsequent runs on the same test host.

Other points of note:
* Removed pre- and post-command hooks
  These used to start up and shut down the Cog server to test
  against. Now that's handled automatically by docker-compose.
* Ports are not mapped to the host machine
  Cog ports (and really, only port 4000) need only be exposed inside the
  docker-compose network, not to the testing host. This prevents clashes
  when multiple docker-compose-based tests are run simultaneously on one
  host.
* The docker-compose Cog server runs from `operable/cog-preview:latest`, unconditionally.
  A future update would be to take advantage of Buildkite's build
  artifacts to be able to export images from the Cog build and use
  _those_ here instead.
* Integration tests can still be run on a workstation without `docker-compose`.