Properly run cog-triggered pipelines
Cog builds now trigger a cogctl pipeline at the end. We now retrieve
the code from the branch of Cog that just built and substitute it for
the `cog-preview` image we would otherwise use.

We can't use Buildkite's built-in Docker Compose support because it
doesn't appear to support layering of compose files, which is what we
need here. As a result, we roll our own support, but using pretty much
the exact same approach that Buildkite takes.

We only bother running the integration tests, because that's what hits
the actual server.

Ideally, we'd also want the "normal" cogctl runs (i.e., not triggered by
cog) to pull from a matching branch (or master), but we can let that
wait until we get the real Dockerfile in cog refactored a bit; right now
the builds for that image can be a bit long, and we can make it shorter.

[skip ci]