Initial Dockerfile
The intention is to create an executable Dockerfile that people could
use to run `cogctl` anywhere (anywhere with Docker, that is).

    docker build -t cogctl .
    docker -v ~/.cogctl:/root/.cogctl run cogctl users
    # etc

In order to make this work, I had to change how `cogctl` writes out a
configuration file. Previously, it would write to a temporary file,
then rename it. This won't work in a Docker environment when the file
being written is bind mounted. The changes would never make it back to
the host system.

Now, we copy the contents of the temporary file into the real file
(instead of renaming), then delete the temporary file.

Incidentally, when we write out a config file, it's a lossy process. Any
comments that existed in the file we read disappear.