Promote master-built images to the cogctl repo
To make our current Cog builds easier, we're going to start pulling the
binaries out of the containers we build them in here in CI. (We're
opting for that right now as opposed to setting up some kind of artifact
repository. This is for expediency and incremental improvement; we'll
get an artifact repository soon.)

When we build on the "master" branch, we'll take the built image (which
we now push to our `cogctl-testing` repository, since there's no
guarantee we'll be running on the same agent that built the image in the
first place) and push it to our `cogctl` repository, tagged with both the platform and "master".

This tag will "float" with whatever the current master branch generates;
it will be replaced the next time the build runs. A `git_commit` Docker
label is added to the image, though, which can be used to determine theh
code it came from.