Add --exact-match to tag detection
Previously, we relied on the output of `git describe --tags`, which
appends information on how many commits away from the last tag the
current commit is. We subsequently processed this to try and determine
if the output referred directly to a tag. Unfortunately, this would
erroneously filter out tags that contained hyphens.

Now, we add `--exact-match` to the invocation, which will return the
name of the tag the current commit lies on, if indeed it does.

Note that this falls apart if more than one tag point to the current
commit, but with how we tag, that's not likely to occur.